# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Aeza Server API
  version: 1.0.0
  description: REST, WebSocket and SSE API for frontend and agents.
servers:
  - url: http://158.160.205.79:8087
    description: Current deployed server
  - url: /
    description: Relative current server
paths:
...
  /api/v1/dashboard/summary:
    get:
      summary: Get summary metrics
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Summary response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
              example:
                totalLogs: 120
                errors: 18
                warnings: 22
                hosts: 4
                services: 6
                aggregatedEvents: 11
                sampledEvents: 4
                burstEvents: 7
  /api/v1/dashboard/timeseries:
    get:
      summary: Get timeseries
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Timeseries response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTimeseries"
              example:
                buckets:
                  - timestamp: "2026-03-21T11:00:00Z"
                    errors: 3
                    warnings: 2
  /api/v1/dashboard/top-hosts:
    get:
      summary: Get top hosts
      description: Returns hosts sorted by observed log volume.
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top hosts response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTopHosts"
              example:
                items:
                  - host: "srv-1"
                    count: 53
                  - host: "srv-2"
                    count: 17
  /api/v1/dashboard/top-services:
    get:
      summary: Get top services
      description: Returns services sorted by observed log volume.
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top services response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTopServices"
              example:
                items:
                  - service: "backend"
                    count: 53
                  - service: "postgres"
                    count: 11
  /api/v1/dashboard/top-patterns:
    get:
      summary: Get top patterns
      description: Returns noisy patterns grouped by fingerprint and message template.
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top patterns response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTopPatterns"
              example:
                items:
                  - fingerprint: "d5c98f0a8be1d4ac"
                    template: "database failed <num>"
                    count: 96
                    sampled: true
                    burstDetected: true
...
components:
  schemas:
...
    DashboardSummary:
      type: object
      properties:
        totalLogs:
          type: integer
          format: int64
        errors:
          type: integer
          format: int64
        warnings:
          type: integer
          format: int64
        hosts:
          type: integer
          format: int64
        services:
          type: integer
          format: int64
        aggregatedEvents:
          type: integer
          format: int64
        sampledEvents:
          type: integer
          format: int64
        burstEvents:
          type: integer
          format: int64
    DashboardTimeseriesBucket:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        errors:
          type: integer
          format: int64
        warnings:
          type: integer
          format: int64
    DashboardTimeseries:
      type: object
      properties:
        buckets:
          type: array
          items:
            $ref: "#/components/schemas/DashboardTimeseriesBucket"
    DashboardTopHostItem:
      type: object
      properties:
        host:
          type: string
        count:
          type: integer
          format: int64
    DashboardTopHosts:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardTopHostItem"
    DashboardTopServiceItem:
      type: object
      properties:
        service:
          type: string
        count:
          type: integer
          format: int64
    DashboardTopServices:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardTopServiceItem"
    DashboardTopPatternItem:
      type: object
      properties:
        fingerprint:
          type: string
        template:
          type: string
        count:
          type: integer
          format: int64
        sampled:
          type: boolean
        burstDetected:
          type: boolean
    DashboardTopPatterns:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardTopPatternItem"
...